requirements.txt
# ---- Web framework
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1

# ---- WSGI server for production
gunicorn==22.0.0
